backend/.requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pydantic
openai
requests
apscheduler
aiofiles
python-multipart
twilio
dateparser
